% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostat3D.R
\name{anisotropy3D}
\alias{anisotropy3D}
\title{Anisotropy ellipsoid}
\usage{
anisotropy3D(maxrange, midrange, minrange = midrange, azimuth = 0,
  dip = 0, rake = 0, radians = F)
}
\arguments{
\item{maxrange, midrange, minrange}{The  three semi-axes of the anisotropy
ellipsoid.}

\item{azimuth, dip, rake}{Orientation of the ellipsoid in geological
coordinates.}

\item{radians}{Are the angles in radians or degrees?}
}
\value{
A 3x3 matrix. Multiplication of a matrix of coordinates by this
matrix yields the transformed coordinates. Multiplication by its inverse
reverses the effect.
}
\description{
Builds a matrix that can be used to rotate and/or strech a set of coordinates
in order to model anisotropy.
}
