% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ara.collar}
\alias{ara.collar}
\title{Drillholes from the Ararangua region}
\format{A data frame with 13 observations and 5 variables:
\describe{
  \item{HOLEID}{Drillhole identification.}
  \item{X, Y, Z}{Collar coordinates.}
  \item{Length}{Total length of hole.}
}}
\source{
\url{http://www.cprm.gov.br/publique/Redes-Institucionais/Rede-de-Bibliotecas---Rede-Ametista/Acervo/Biblioteca-Virtual-1327.html}
}
\usage{
ara.collar
}
\description{
A dataset containing drillhole data from a coal prospecting project in the
Ararangua region, southern Brazil.
}
\keyword{datasets}
