% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ara.lito}
\alias{ara.lito}
\title{Drillholes from the Ararangua region}
\format{A data frame with 507 observations and 6 variables:
\describe{
  \item{HOLEID}{Drillhole identification.}
  \item{FROM, TO}{Hole segment delimiters.}
  \item{LITO}{Lithological description.}
  \item{Form}{Geological formation.}
  \item{Camada}{Names of the coal seams. Rock layers that are not coal have
                no names.}
}}
\source{
\url{http://www.cprm.gov.br/publique/Redes-Institucionais/Rede-de-Bibliotecas---Rede-Ametista/Acervo/Biblioteca-Virtual-1327.html}
}
\usage{
ara.lito
}
\description{
A dataset containing drillhole data from a coal prospecting project in the
Ararangua region, southern Brazil.
}
\keyword{datasets}
