% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/points3DDataFrame.R
\docType{methods}
\name{Bind}
\alias{Bind}
\alias{Bind,points3DDataFrame,points3DDataFrame-method}
\title{Bind}
\usage{
Bind(x, y, ...)

\S4method{Bind}{points3DDataFrame,points3DDataFrame}(x, y)
}
\arguments{
\item{x, y}{\code{spatial3DDataFrame} objects.}
}
\value{
The resulting object's class will depend of the classes of the
parent objects.
}
\description{
Binds two 3D spatial objects together.
}
\details{
The two objects' coordinates are binded as in the \code{rbind}
function. If the object's column names are not the same, the resulting
object will have all columns from both, with the appropriate entries filled
with \code{NA}. Note that if the parent objects have coordinates in common,
the resulting object will have duplicated coordinates.
}
