% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks3DDataFrame.R
\docType{methods}
\name{blocks3DDataFrame-init}
\alias{blocks3DDataFrame-init}
\title{3D block model}
\usage{
\S4method{initialize}{blocks3DDataFrame}(.Object, corner, number, size,
  discretization = c(4, 4, 4), fields = ".dummy")
}
\arguments{
\item{corner}{A length 3 vector containing the smallest coordinate in each
direction.}

\item{number}{A length 3 vector containing the desired number of blocks in
each direction.}

\item{size}{A length 3 vector with the block dimensions in each direction.}

\item{discretization}{A length 3 vector indicating in how many parts to
"break" each block during the computations.}

\item{fields}{Column names to populate the grid. The columns are
automatically filled with \code{NA}s.}
}
\description{
A 3D grid with adjacent blocks. Extends the \code{grid3DDataFrame}
class.
}
