% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks3DDataFrame.R
\name{blocks3DDataFrame}
\alias{blocks3DDataFrame}
\alias{blocks3DDataFrame}
\title{3D block model}
\usage{
blocks3DDataFrame(.Object, corner, number, size, discretization = c(4, 4, 4),
  fields = ".dummy")

blocks3DDataFrame(.Object, corner, number, size, discretization = c(4, 4, 4),
  fields = ".dummy")
}
\arguments{
\item{corner}{A length 3 vector containing the smallest coordinate in each
direction.}

\item{number}{A length 3 vector containing the desired number of blocks in
each direction.}

\item{size}{A length 3 vector with the block dimensions in each direction.}

\item{discretization}{A length 3 vector indicating in how many parts to
"break" each block during the computations.}

\item{fields}{Column names to populate the grid. The columns are
automatically filled with \code{NA}s.}
}
\description{
A 3D grid with adjacent blocks. Extends the \code{grid3DDataFrame}
class.

A 3D grid with adjacent blocks. Extends the \code{grid3DDataFrame}
class.
}
\details{
Subsetting an object of this class will coerce the output to a
\code{points3DDataFrame} object.
}
\section{Slots}{

\describe{
\item{\code{size}}{The block size in the three dimensions.}

\item{\code{discretization}}{The number of subdivisions for computation of
covariances.}
}}

\seealso{
\code{\link{blocks3DDataFrame-init}}
}
