% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/spatial3DDataFrame.R
\docType{methods}
\name{BoundingBox}
\alias{BoundingBox}
\alias{BoundingBox,spatial3DDataFrame-method}
\title{Bounding box}
\usage{
BoundingBox(object, ...)

\S4method{BoundingBox}{spatial3DDataFrame}(object)
}
\arguments{
\item{object}{The object from which to obtain the bounding box.}
}
\value{
A 2x3 matrix containing the lower and upper bounds of the object's
coordinates.
}
\description{
Returns a \code{spatial3DDataFrame} object's, bounding box, the smallest
rectangular parallelepiped that contains all the data.
}
