% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/geostat3D.R
\docType{methods}
\name{CovarianceMatrix}
\alias{CovarianceMatrix}
\alias{CovarianceMatrixD1}
\alias{CovarianceMatrixD2}
\alias{CovarianceMatrix,lines3DDataFrame,lines3DDataFrame-method}
\alias{CovarianceMatrix,lines3DDataFrame,points3DDataFrame-method}
\alias{CovarianceMatrix,points3DDataFrame,lines3DDataFrame-method}
\alias{CovarianceMatrix,points3DDataFrame,points3DDataFrame-method}
\alias{CovarianceMatrixD1,points3DDataFrame,points3DDataFrame-method}
\alias{CovarianceMatrixD2,points3DDataFrame-method}
\title{Covariance matrix of spatial data and/or its derivatives.}
\usage{
CovarianceMatrix(x, y, ...)

CovarianceMatrixD1(x, tangents, ...)

CovarianceMatrixD2(tangents, ...)

\S4method{CovarianceMatrix}{lines3DDataFrame,lines3DDataFrame}(x, y, model,
  covariance = T, parts = 10)

\S4method{CovarianceMatrix}{lines3DDataFrame,points3DDataFrame}(x, y, model,
  covariance = T, parts = 10)

\S4method{CovarianceMatrix}{points3DDataFrame,lines3DDataFrame}(x, y, model,
  covariance = T, parts = 10)

\S4method{CovarianceMatrix}{points3DDataFrame,points3DDataFrame}(x, y, model,
  covariance = T)

\S4method{CovarianceMatrixD1}{points3DDataFrame,points3DDataFrame}(x, tangents,
  model, covariance = T)

\S4method{CovarianceMatrixD2}{points3DDataFrame}(tangents, model,
  covariance = T)
}
\arguments{
\item{x, y}{3D spatial objects.}

\item{tangents}{A \code{points3DDataFrame} containing directional data, most
likely generated with the \code{GetPlaneDirections()} method.}

\item{model}{A \code{covarianceStructure3D} object representing the spatial
continuity of the variable of interest, or a \code{list} containing multiple
such objects.}

\item{covariance}{Should the matrix be calculated as a covariance matrix
(default) or a variance (gamma) matrix?}

\item{parts}{The number of parts in which to "break" line segments.}
}
\value{
A covariance matrix with dimensions depending of the method called:
\describe{
  \item{\code{CovarianceMatrix()}}{ \code{nrow(x)} rows and \code{nrow(y)}
  columns}
  \item{\code{CovarianceMatrixD1()}}{ \code{nrow(x)} rows and
  \code{nrow(tangents)} columns}
  \item{\code{CovarianceMatrixD2()}}{ \code{nrow(tangents)} rows and
  \code{nrow(tangents)} columns}
}
}
\description{
Calculates the covariance matrix of a spatially distributed value and/or its
derivative.
}
\details{
If \code{covariance = F} the resulting matrix is given in variogram
form, with zeroes in the main diagonal and values that increase with the
distance between data locations.

If the spatial objects are not made of punctual data, the covariance
is calcuated with a "brute force" approach, by discretizing each data
element in points and calculating an average covariance.

The \code{CovarianceMatrixD1()} method calculates the covariances between
the variable of interest located in \code{x} and its derivatives located in
\code{tangents}. The \code{CovarianceMatrixD2()} calculates the covariances
between the derivatives themselves.
}
\seealso{
\code{\link{covarianceStructure3D-class}},
\code{\link{GetPlaneDirections}}
}
