% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covarianceModel3D.R
\name{covarianceModel3D-class}
\alias{covarianceModel3D-class}
\alias{covarianceModel3D}
\title{Covariance model}
\usage{
covarianceModel3D(nugget, structures, nugget.dir = 0)
}
\description{
Representation of a spatial continuity model.
}
\section{Slots}{

\describe{
\item{\code{nugget}}{The nugget effect or noise variance.}

\item{\code{nugget.dir}}{The nugget effect for directional data.}

\item{\code{structures}}{A \code{list} with one or more
\code{covarianceStructure3D} objects}
}}

\seealso{
\code{\link{covarianceStructure3D-init}},
\code{\link{CovarianceMatrix}}
}
