% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covarianceModel3D.R
\name{covarianceModel3D-init}
\alias{covarianceModel3D-init}
\alias{covarianceModel3D}
\title{Covariance model}
\usage{
covarianceModel3D(nugget, structures, nugget.dir = 0)
}
\arguments{
\item{nugget}{The nugget effect or noise variance.}

\item{structures}{Objects of the \code{covarianceStructure3D} class
describing the desired covariance structure. A single object or multiple,
concatenated objects.}

\item{nugget.dir}{The nugget effect for directional data.}
}
\description{
Representation of a spatial continuity model.
}
\seealso{
\code{\link{covarianceModel3D-class}},
\code{\link{covarianceStructure3D-class}}
}
