% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covarianceModel3D.R
\name{covarianceModel3D}
\alias{covarianceModel3D}
\alias{covarianceModel3D}
\title{Covariance model}
\usage{
covarianceModel3D(nugget, structures, nugget.dir = 0)

covarianceModel3D(nugget, structures, nugget.dir = 0)
}
\arguments{
\item{nugget}{The nugget effect or noise variance.}

\item{structures}{Objects of the \code{covarianceStructure3D} class
describing the desired covariance structure. A single object or multiple,
concatenated objects.}

\item{nugget.dir}{The nugget effect for directional data.}
}
\description{
Representation of a spatial continuity model.

Representation of a spatial continuity model.
}
\section{Slots}{

\describe{
\item{\code{nugget}}{The nugget effect or noise variance.}

\item{\code{nugget.dir}}{The nugget effect for directional data.}

\item{\code{structures}}{A \code{list} with one or more
\code{covarianceStructure3D} objects}

\item{\code{total.var}}{The total variance of the covariance structures,
excluding the nugget effect.}
}}

\seealso{
\code{\link{covarianceStructure3D-init}},
\code{\link{CovarianceMatrix}}

\code{\link{covarianceStructure3D-class}}
}
