% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covarianceStructure3D.R
\name{covarianceStructure3D-class}
\alias{covarianceStructure3D-class}
\alias{covarianceStructure3D}
\title{Covariance structure}
\usage{
covarianceStructure3D(type = c("gaussian", "exponential", "spherical",
  "cubic", "matern1", "matern2", "cauchy"), contribution, maxrange,
  midrange = maxrange, minrange = midrange, azimuth = 0, dip = 0,
  rake = 0, power = 1)
}
\description{
Representation of one structure in a spatial continuity model.
}
\section{Slots}{

\describe{
\item{\code{type}}{The type of covariance function represented.}

\item{\code{contribution}}{The variance associated with this structure.}

\item{\code{maxrange,midrange,minrange}}{The structure's range in three orthogonal
directions, or the three semi-axes of the covariance ellipsoid.}

\item{\code{azimuth,dip,rake}}{Orientation of the covariance ellipsoid in geological
coordinates.}

\item{\code{power}}{The power parameter for the covariance functions that feature
one.}

\item{\code{anis}}{An anisotropy matrix.}

\item{\code{covfun,covd1,covd2,varfun,vard1,vard2}}{Functions to calculate the
covariance and variogram matrices between a variable of interest and/or its
derivatives.}
}}

\seealso{
\code{\link{covarianceStructure3D-init}},
\code{\link{CovarianceMatrix}}
}
