% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covarianceStructure3D.R
\name{covarianceStructure3D}
\alias{covarianceStructure3D}
\alias{covarianceStructure3D}
\title{Covariance structure}
\usage{
covarianceStructure3D(type = c("gaussian", "exponential", "spherical",
  "cubic", "matern1", "matern2", "cauchy"), contribution, maxrange,
  midrange = maxrange, minrange = midrange, azimuth = 0, dip = 0,
  rake = 0, power = 1)

covarianceStructure3D(type = c("gaussian", "exponential", "spherical",
  "cubic", "matern1", "matern2", "cauchy"), contribution, maxrange,
  midrange = maxrange, minrange = midrange, azimuth = 0, dip = 0,
  rake = 0, power = 1)
}
\arguments{
\item{type}{The type of covariance function represented (see below).}

\item{contribution}{The variance associated with this structure.}

\item{maxrange, midrange, minrange}{The structure's range in three orthogonal
directions, or the three semi-axes of the covariance ellipsoid.}

\item{azimuth, dip, rake}{Orientation of the covariance ellipsoid in geological
coordinates.}

\item{power}{The power parameter for the covariance functions that feature
one.}
}
\description{
Representation of one structure in a spatial continuity model.

Representation of one structure in a spatial continuity model.
}
\section{Slots}{

\describe{
\item{\code{type}}{The type of covariance function represented.}

\item{\code{contribution}}{The variance associated with this structure.}

\item{\code{maxrange,midrange,minrange}}{The structure's range in three orthogonal
directions, or the three semi-axes of the covariance ellipsoid.}

\item{\code{azimuth,dip,rake}}{Orientation of the covariance ellipsoid in geological
coordinates.}

\item{\code{power}}{The power parameter for the covariance functions that feature
one.}

\item{\code{anis}}{An anisotropy matrix.}

\item{\code{covfun,covd1,covd2,varfun,vard1,vard2}}{Functions to calculate the
covariance and variogram matrices between a variable of interest and/or its
derivatives.}
}}

\section{Covariance functions}{


The following functions are available for the \code{type} parameter. In the
equations below, \eqn{d} is the anisotropy-corrected distance between two points
and \eqn{p} is the \code{power} parameter. Functions marked with an asterisk
support modeling with structural data.

\describe{
  \item{\code{gaussian}*}{\deqn{\exp{-3d^2}}{exp(-3*d^2)}}
  \item{\code{exponential}}{\deqn{\exp{-3d}}{exp(-3*d)}}
  \item{\code{spherical}}{\deqn{1 - 1.5 d + 0.5 d^3, 0 \leq d \leq 1}{1 - 1.5*d + 0.5*d^3, 0 <= d <= 1}}
  \item{\code{cubic}*}{\deqn{1 - 7d^2 + \frac{35}{4}d^3 - \frac{7}{2}d^5 + \frac{3}{4}d^7, 0 \leq d \leq 1}{1 - 7*d^2 + 35/4*d^3 - 7/2*d^5 + 3/4*d^7, 0 <= d <= 1}}
  \item{\code{matern1}*}{\deqn{(1 + 5d)\exp{-5d}}{(1 + 5*d) * exp(-5*d)}}
  \item{\code{matern2}*}{\deqn{(1 + 6d + 12d^2)\exp{-6d}}{(1 + 6*d + 12*d^2) * exp(-6*d)}}
  \item{\code{cauchy}*}{\deqn{(1 + d^2)^{-p}, p > 0}{(1 + d^2)^(-p), p > 0}}
}
}

\seealso{
\code{\link{covarianceStructure3D-init}},
\code{\link{CovarianceMatrix}}

\code{\link{covarianceStructure3D-class}}
}
