% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covarianceStructure3DNonStat.R
\name{covarianceStructure3DNonStat-class}
\alias{covarianceStructure3DNonStat-class}
\alias{covarianceStructure3DNonStat}
\title{Non stationary covariance structure}
\usage{
covarianceStructure3DNonStat(type = c("gaussian", "exponential", "spherical",
  "cubic", "matern1", "matern2", "cauchy"), contribution, maxrange,
  midrange = maxrange, minrange = midrange, azimuth = 0, dip = 0,
  rake = 0, power = 1, control_points)
}
\description{
Representation of one structure in a non stationary spatial continuity model.
}
\section{Slots}{

\describe{
\item{\code{type}}{The type of covariance function represented.}

\item{\code{contribution}}{The variance associated with this structure.}

\item{\code{maxrange,midrange,minrange}}{The structure's range in three orthogonal
directions, or the three semi-axes of the covariance ellipsoid.}

\item{\code{azimuth,dip,rake}}{Orientation of the covariance ellipsoid in geological
coordinates.}

\item{\code{power}}{The power parameter for the covariance functions that feature
one.}

\item{\code{covfun,covd1,covd2}}{Functions to calculate the
covariance matrices between a variable of interest and/or its
derivatives.}
}}

\seealso{
\code{\link{covarianceStructure3DNonStat-init}},
\code{\link{CovarianceMatrix}}
}
