% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covarianceStructure3DNonStat.R
\name{covarianceStructure3DNonStat-init}
\alias{covarianceStructure3DNonStat-init}
\alias{covarianceStructure3DNonStat}
\title{Non stationary covariance structure}
\usage{
covarianceStructure3DNonStat(type = c("gaussian", "exponential", "spherical",
  "cubic", "matern1", "matern2", "cauchy"), contribution, maxrange,
  midrange = maxrange, minrange = midrange, azimuth = 0, dip = 0,
  rake = 0, power = 1, control_points)
}
\arguments{
\item{type}{The type of covariance function represented (see below).}

\item{contribution}{The variance associated with this structure.}

\item{maxrange, midrange, minrange}{The structure's range in three orthogonal
directions, or the three semi-axes of the covariance ellipsoid.}

\item{azimuth, dip, rake}{Orientation of the covariance ellipsoid in geological
coordinates.}

\item{power}{The power parameter for the covariance functions that feature
one.}

\item{control_points}{A \code{points3DDataFrame} object with the coordinates
of the non stationary features.}
}
\description{
Representation of one structure in a non stationary spatial continuity model.
}
\details{
All parameters except \code{power} are non stationary, each
modeled by its own (stationary, noiseless) \code{GP} and using the control points as
data. The GPs are initialized with the values provided as data, and by
default predict a constant value everywhere. Use the \code{Fit} method of the
main GP to determine the best value for each parameter at the control points
and the stationary covariance model for each.
}
\section{Covariance functions}{


The following functions are available for the \code{type} parameter. In the
equations below, \eqn{d} is the anisotropy-corrected distance between two points
and \eqn{p} is the \code{power} parameter. Functions marked with an asterisk
support modeling with directional data.

\describe{
  \item{\code{gaussian}*}{\deqn{\exp{-3d^2}}{exp(-3*d^2)}}
  \item{\code{exponential}}{\deqn{\exp{-3d}}{exp(-3*d)}}
  \item{\code{spherical}}{\deqn{1 - 1.5 d + 0.5 d^3, 0 \leq d \leq 1}{1 - 1.5*d + 0.5*d^3, 0 <= d <= 1}}
  \item{\code{cubic}*}{\deqn{1 - 7d^2 + \frac{35}{4}d^3 - \frac{7}{2}d^5 + \frac{3}{4}d^7, 0 \leq d \leq 1}{1 - 7*d^2 + 35/4*d^3 - 7/2*d^5 + 3/4*d^7, 0 <= d <= 1}}
  \item{\code{matern1}*}{\deqn{(1 + 5d)\exp{-5d}}{(1 + 5*d) * exp(-5*d)}}
  \item{\code{matern2}*}{\deqn{(1 + 6d + 12d^2)\exp{-6d}}{(1 + 6*d + 12*d^2) * exp(-6*d)}}
  \item{\code{cauchy}*}{\deqn{(1 + d^2)^{-p}, p > 0}{(1 + d^2)^(-p), p > 0}}
}
}

\seealso{
\code{\link{covarianceStructure3DNonStat-class}},
\code{\link{covarianceModel3D-class}}
}
