% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directions3DDataFrame.R
\name{directions3DDataFrame-class}
\alias{directions3DDataFrame-class}
\alias{directions3DDataFrame}
\title{directions3DDataFrame-class}
\usage{
directions3DDataFrame(coords, df, directions)
}
\section{Slots}{

\describe{
\item{\code{coords}}{A \code{list} containing the coordinates for each data unit.}

\item{\code{data}}{A \code{data.frame} with each data unit's attributes.}

\item{\code{bbox}}{A \code{matrix} containing the coordinates of two opposing
edges of the object's bounding box.}

\item{\code{directions}}{A \code{matrix} containing coordinates of unit vectors
representing directions.}
}}

\seealso{
\code{\link{directions3DDataFrame-init}},
\code{\link{spatial3DDataFrame-class}},
\code{\link{points3DDataFrame-init}}
}
