% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directions3DDataFrame.R
\name{directions3DDataFrame}
\alias{directions3DDataFrame}
\title{3D Directional data}
\usage{
directions3DDataFrame(coords, df, directions)
}
\arguments{
\item{coords}{A list, matrix, or data frame containing the 3D coordinates
of the points.}

\item{df}{A data frame with the points' attributes}

\item{directions}{A matrix, or data frame containing the 3D coordinates
of the directional vectors.}
}
\description{
3D Directional data
}
\details{
If \code{coords} or \code{directions} have less than 3
columns, the missing coordinates are given a value of 0. The directions are
rescaled to unit length.
}
\seealso{
\code{\link{directions3DDataFrame-class}},
\code{\link{points3DDataFrame-class}}
}
