% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/directions3DDataFrame.R
\docType{methods}
\name{DrawDirections}
\alias{DrawDirections}
\alias{DrawDirections,directions3DDataFrame-method}
\title{Visualization of structural data}
\usage{
DrawDirections(object, ...)

\S4method{DrawDirections}{directions3DDataFrame}(object, size, col = "yellow",
  as = c("direction", "arrow"))
}
\arguments{
\item{size}{The length of the cylinders.}

\item{col}{The color of the discs. Either a single value or a vector
matching the object's number of rows.}

\item{as}{How to plot the data. \code{"direction"} is useful for
structural data, while \code{"arrow"} can be used for for velocities and
derivative data in a more strict sense.}
}
\description{
Draws thin cylinders representing the orientation of structural lines or
directional data in general.
}
\details{
\code{col} must contain colors in hexadecimal format, such as
\code{"#804DB3"}, or valid color names.
}
