% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/lines3DDataFrame.R
\docType{methods}
\name{DrawDrillholes}
\alias{DrawDrillholes}
\alias{DrawDrillholes,lines3DDataFrame-method}
\title{Visualization of drillhole data}
\usage{
DrawDrillholes(object, ...)

\S4method{DrawDrillholes}{lines3DDataFrame}(object, by, values, col, size,
  col.default = "white")
}
\arguments{
\item{by}{Name of the column to show.}

\item{values, col}{If \code{by} is a categorical variable, the values to show
and the corresponding color. If \code{by} is a continuous variable, a numeric
vector with the breakpoints and the corresponding colors.}

\item{size}{The diameter of the cylinders. Either a single value or a vector
matching the object's number of rows.}

\item{col.default}{Color to be used for \code{NA} values and values outside
the range provided.}
}
\description{
Adds drillhole information to the current \code{rgl} window as colored
cylinders.
}
\details{
For categorical variables, each entry in \code{values} is matched to
the corresponding entry in \code{col}. For continuous variables,
\code{values} is a numeric vector with breakpoints, each of which is matched
to the entries in \code{col}. The actual values are interpolated to generate
a unique color to each value. \code{col} must be a character vector
containing colors in hexadecimal format, such as \code{"#804DB3"}, or valid
color names.
}
