% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/lines3DDataFrame.R
\docType{methods}
\name{DrawHoleID}
\alias{DrawHoleID}
\alias{DrawHoleID,lines3DDataFrame-method}
\title{Visualization of drillhole data}
\usage{
DrawHoleID(object, ...)

\S4method{DrawHoleID}{lines3DDataFrame}(object, cex = 1)
}
\arguments{
\item{cex}{The character expansion factor for the labels.}
}
\description{
Draws the hole identifications in the current \code{rgl} window.
}
