% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/points3DDataFrame.R
\docType{methods}
\name{DrawPoints}
\alias{DrawPoints}
\alias{DrawPoints,points3DDataFrame-method}
\title{Visualization of point data}
\usage{
DrawPoints(object, ...)

\S4method{DrawPoints}{points3DDataFrame}(object, by, values, col, size,
  alpha = 1, col.default = "white", as = c("spheres", "points"))
}
\arguments{
\item{by}{Name of the column to show.}

\item{values, col}{If \code{by} is a categorical variable, the values to show
and the corresponding color. If \code{by} is a continuous variable, a numeric
vector with the breakpoints and the corresponding colors.}

\item{size}{The diameter of the spheres. Either a single value or a vector
matching the object's number of rows.}

\item{alpha}{Opacity factor.}

\item{col.default}{Color to be used for \code{NA} values and values outside
the range provided.}

\item{as}{How to plot the points. For \code{"points"}, only the color
information is used but the visualization becomes much lighter in terms of
machine resources.}
}
\description{
Adds point information to the current \code{rgl} window in the form of
colored spheres or points.
}
\details{
For categorical variables, each entry in \code{values} is matched to
the corresponding entry in \code{col}. For continuous variables,
\code{values} is a numeric vector with breakpoints, each of which is matched
to the entries in \code{col}. The actual values are interpolated to generate
a unique color to each value. \code{col} must be a character vector
containing colors in hexadecimal format, such as \code{"#804DB3"}, or valid
color names.
}
