% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/grid3DDataFrame.R
\docType{methods}
\name{DrawSection}
\alias{DrawSection}
\alias{DrawSection,grid3DDataFrame-method}
\title{DrawSection}
\usage{
DrawSection(object, ...)

\S4method{DrawSection}{grid3DDataFrame}(object, by, values, col,
  col.default = "white", x = NULL, y = NULL, z = NULL)
}
\arguments{
\item{by}{Name of the column to show.}

\item{values, col}{If \code{by} is a categorical variable, the values to show
and the corresponding color. If \code{by} is a continuous variable, a numeric
vector with the breakpoints and the corresponding colors.}

\item{col.default}{Color to be used for \code{NA} values and values outside
the range provided.}

\item{x, y, z}{The coordinate to be kept fixed in order to draw the section (
will be matched to the closest corresponding coordinate in \code{object}).
Only one of these must be provided.}
}
\description{
Draws a 2D section of a regularly spaced 3D object in the current \code{rgl}
window.
}
\details{
The section is drawn by keeping one coordinate fixed. For example,
if \code{x=5} is given, the section will be drawn in the yz plane.
}
