% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/points3DDataFrame.R
\docType{methods}
\name{DrawTangentLines}
\alias{DrawTangentLines}
\alias{DrawTangentLines,points3DDataFrame-method}
\title{Visualization of structural data}
\usage{
DrawTangentLines(object, ...)

\S4method{DrawTangentLines}{points3DDataFrame}(object, size, dX = "dX",
  dY = "dY", dZ = "dZ", col = "yellow")
}
\arguments{
\item{size}{The length of the cylinders.}

\item{dX, dY, dZ}{Names of the columns that contain the structural lines'
directions.}

\item{col}{The color of the discs. Either a single value or a vector
matching the object's number of rows.}
}
\description{
Draws thin cylinders representing the orientation of structural lines.
}
\details{
\code{col} must contain colors in hexadecimal format, such as
\code{"#804DB3"}, or valid color names.

It is assumed that \code{dX}, \code{dY}, and \code{dZ} are coordinates of
a unit vector.
}
