% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/points3DDataFrame.R
\docType{methods}
\name{DrawTangentPlanes}
\alias{DrawTangentPlanes}
\alias{DrawTangentPlanes,points3DDataFrame-method}
\title{Visualization of structural data}
\usage{
DrawTangentPlanes(object, ...)

\S4method{DrawTangentPlanes}{points3DDataFrame}(object, size, dip = "Dip",
  strike = "Strike", col = "yellow")
}
\arguments{
\item{size}{The diameter of the discs.}

\item{dip, strike}{Names of the columns that contain the structural
information.}

\item{col}{The color of the discs. Either a single value or a vector
matching the object's number of rows.}
}
\description{
Draws discs representing the orientation of structural planes.
}
\details{
\code{col} must contain colors in hexadecimal format, such as
\code{"#804DB3"}, or valid color names.

It is assumed that the values in the \code{dip} and \code{strike} columns
are in degrees.
}
