% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/lines3DDataFrame.R
\docType{methods}
\name{GetContacts}
\alias{GetContacts}
\alias{GetContacts,lines3DDataFrame-method}
\title{GetContacts}
\usage{
GetContacts(object, ...)

\S4method{GetContacts}{lines3DDataFrame}(object, by)
}
\arguments{
\item{by}{A column name indicating which attribute to use to find contacts.}
}
\value{
A \code{points3DDataFrame} object with three attributes:
\describe{
  \item{HOLEID}{The identification of the hole from which the point was
  extracted.}
  \item{.up}{The value of the \code{by} variable in the segment above the
  point}
  \item{.down}{The value of the \code{by} variable in the segment below the
  point}
}
}
\description{
Finds the points of contact between different geological units or other
categorical variable.
}
\details{
This method extracts the points that lie in the border between any
two contiguous line segments with different values of the \code{by} column.
It is assumed that \code{by} represents a categorical variable.
}
