% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/points3DDataFrame.R,
%   R/lines3DDataFrame.R
\docType{methods}
\name{GetCoords}
\alias{GetCoords}
\alias{GetCoords,points3DDataFrame-method}
\alias{GetCoords,lines3DDataFrame-method}
\title{\code{spatial3DDataFrame} objects' coordinates}
\usage{
GetCoords(object, ...)

\S4method{GetCoords}{points3DDataFrame}(object, as = c("list", "data.frame",
  "matrix"))

\S4method{GetCoords}{lines3DDataFrame}(object, as = c("list", "data.frame",
  "matrix"))
}
\arguments{
\item{object}{The object to get coordinates from.}

\item{as}{The return format. One of \code{"list"}, \code{"matrix"}, or
\code{"data.frame"}.}
}
\value{
An object according to the \code{as} parameter containing the
object's coordinates.
}
\description{
Returns the spatial coordinates of a \code{spatial3DDataFrame} object.
}
