% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/spatial3DDataFrame.R
\docType{methods}
\name{GetData}
\alias{GetData}
\alias{GetData,spatial3DDataFrame-method}
\title{\code{spatial3DDataFrame} objects' data}
\usage{
GetData(object, ...)

\S4method{GetData}{spatial3DDataFrame}(object)
}
\arguments{
\item{object}{The object to get the data from.}
}
\value{
A data frame with the object's data.
}
\description{
Returns the attributes of a \code{spatial3DDataFrame} object, represented
as a \code{data.frame}.
}
