% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/lines3DDataFrame.R
\docType{methods}
\name{GetLength}
\alias{GetLength}
\alias{GetLength,lines3DDataFrame-method}
\title{Drillhole segments' lengths}
\usage{
GetLength(object, ...)

\S4method{GetLength}{lines3DDataFrame}(object)
}
\arguments{
\item{object}{The object to get the data from.}
}
\value{
A numeric vector with the segments' lengths.
}
\description{
Returns the length of each segment of a \code{lines3DDataFrame} object.
}
