% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/spatial3DDataFrame.R
\docType{methods}
\name{GetLineDirections}
\alias{GetLineDirections}
\alias{GetLineDirections,spatial3DDataFrame-method}
\title{Structural lines' directions}
\usage{
GetLineDirections(object, ...)

\S4method{GetLineDirections}{spatial3DDataFrame}(object, dip = "Dip",
  azimuth = "Azimuth")
}
\arguments{
\item{object}{A \code{spatial3DDataFrame} object.}

\item{dip, azimuth}{Names of columns containing dip and azimuth information.}
}
\value{
A \code{directions3DDataFrame} containing the
coordinates of the vectors that define a structural
line. The coordinates are normalized to unit length.
}
\description{
Calculates the coordinates of vectors corresponding to structural directions.
}
\details{
Dip and azimuth values are assumed to be in degrees.
}
\seealso{
\code{\link{GetPlaneDirections}}, \code{\link{GetNormals}}
}
