% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/covarianceModel3D.R,
%   R/covarianceStructure3DNonStat.R
\docType{methods}
\name{GetNonStatParams}
\alias{GetNonStatParams}
\alias{GetNonStatParams,covarianceModel3D-method}
\alias{GetNonStatParams,covarianceStructure3DNonStat-method}
\title{Non stationary model parameters}
\usage{
GetNonStatParams(object, ...)

\S4method{GetNonStatParams}{covarianceModel3D}(object, target)

\S4method{GetNonStatParams}{covarianceStructure3DNonStat}(object, target)
}
\arguments{
\item{object}{A \code{covarianceModel3Dobject}.}

\item{target}{A 3D data frame object with the target locations.}
}
\value{
An object with the same class as \code{target}, containing the
parameters (contribution, ranges, etc.).
}
\description{
Returns a non stationary covariance model's parameters at the specified
locations.
}
