% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/spatial3DDataFrame.R
\docType{methods}
\name{GetNormals}
\alias{GetNormals}
\alias{GetNormals,spatial3DDataFrame-method}
\title{Normals to structural planes}
\usage{
GetNormals(object, ...)

\S4method{GetNormals}{spatial3DDataFrame}(object, dip = "Dip",
  strike = "Strike")
}
\arguments{
\item{object}{A \code{spatial3DDataFrame} object.}

\item{dip, strike}{Names of columns containing dip and strike information.}
}
\value{
An object of the same class as \code{object} containing the
coordinates of the normals to the specified planes (nX, nY, nZ). The
coordinates are normalized to unit length.
}
\description{
Calculates the vector normal to a structural plane.
}
\details{
Dip and strike values are assumed to be in degrees.
}
