% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/spatial3DDataFrame.R
\docType{methods}
\name{GetPlaneDirections}
\alias{GetPlaneDirections}
\alias{GetPlaneDirections,spatial3DDataFrame-method}
\title{Structural planes' directions}
\usage{
GetPlaneDirections(object, ...)

\S4method{GetPlaneDirections}{spatial3DDataFrame}(object, dip = "Dip",
  strike = "Strike")
}
\arguments{
\item{object}{A \code{spatial3DDataFrame} object.}

\item{dip, strike}{Names of columns containing dip and strike information.}
}
\value{
An object of the same class as \code{object} (with twice the number
of rows) containing the coordinates of the vectors that define a structural
plane (one parallel to dip and the other parallel to strike). The
coordinates are normalized to unit length.
}
\description{
Calculates structural planes' dip and strike vectors.
}
\details{
Dip and strike values are assumed to be in degrees.
}
