% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/covarianceModel3D.R
\docType{methods}
\name{GetPriorVariance}
\alias{GetPriorVariance}
\alias{GetPriorVariance,covarianceModel3D-method}
\title{Covariance model variance}
\usage{
GetPriorVariance(object, ...)

\S4method{GetPriorVariance}{covarianceModel3D}(object, target)
}
\arguments{
\item{object}{A \code{covarianceModel3Dobject}.}

\item{target}{A 3D data frame object with the target locations.}
}
\value{
A vector containing the variance at the specified locations.
}
\description{
Returns a covariance model's prior variance at the specified locations.
}
