% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GP.R
\name{GP-class}
\alias{GP-class}
\alias{GP}
\title{Gaussian Process}
\usage{
GP(data, model, value, mean = NULL, trend = NULL, weights = NULL,
  force.interp = numeric(), reg.v = 1e-09, tangents = NULL,
  reg.t = 1e-12, nugget.t = 0)
}
\description{
Implementation of the Gaussian Process model for 3D spatial interpolation.
}
\section{Slots}{

\describe{
\item{\code{data}}{A \code{spatial3DDataFrame} object containing the necessary data.}

\item{\code{tangents}}{A \code{directions3DDataFrame} object containing structural
geology data. Most likely generated with the \code{GetLineDirections()}
or \code{GetPlaneDirections()} method.}

\item{\code{model}}{The covariance model. A \code{list} containing one or more
\code{covarianceStructure3D} objects.}

\item{\code{mean}}{The global mean. Irrelevant if a trend is used.}

\item{\code{trend}}{The model's trend component. A formula in character format.}

\item{\code{beta}}{The trend coefficients.}

\item{\code{likelihood}}{The model's log-likelihood given the data.}

\item{\code{pre_comp}}{A \code{list} containing pre-computed values to speed up
predictions.}
}}

\references{
Rasmussen CE, Williams CKI. Gaussian processes for machine
learning. Cambridge, Massachusetts: MIT Press; 2006.
doi:10.1142/S0129065704001899.
}
\seealso{
\code{\link{GP-init}}, \code{\link{SPGP-class}},
\code{\link{GP_geomod-class}}
}
