% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GP_geomod.R
\name{GP_geomod-class}
\alias{GP_geomod-class}
\alias{GP_geomod}
\title{Gaussian Process for Implicit Geological Modeling}
\usage{
GP_geomod(data, value1, value2 = value1, model, tangents = NULL,
  enforce.contacts = T, reg.v = 1e-09, reg.t = 1e-09)
}
\description{
Implicit geological using Gaussian Processes and compositional relations
among the geological classes.
}
\section{Slots}{

\describe{
\item{\code{GPs}}{A \code{list} containing one \code{GP} object per class.}

\item{\code{params}}{A \code{list} containing modeling parameters such as
regularization factors.}
}}

\references{
Gonçalves IG, Kumaira S, Guadagnin F.
A machine learning approach to the potential-field method for implicit
modeling of geological structures. Comput Geosci 2017;103:173–82.
doi:10.1016/j.cageo.2017.03.015.
}
\seealso{
\code{\link{GP_geomod-init}}, \code{\link{GP-class}}
}
