% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid3DDataFrame.R
\docType{class}
\name{grid3DDataFrame-class}
\alias{grid3DDataFrame-class}
\alias{grid3DDataFrame}
\title{3D regular point grid}
\description{
A 3D grid with equally spaced points. Extends the \code{points3DDataFrame}
class.
}
\details{
Subsetting an object of this class will coerce the output to a
\code{points3DDataFrame} object.
}
\section{Slots}{

\describe{
\item{\code{dims}}{The number of grid points in each direction.}
}}

\seealso{
\code{\link{grid3DDataFrame-init}}
}
