% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid3DDataFrame.R
\name{grid3DDataFrame}
\alias{grid3DDataFrame}
\alias{grid3DDataFrame}
\title{3D regular point grid}
\usage{
grid3DDataFrame(gridx, gridy, gridz, fields = ".dummy")

grid3DDataFrame(gridx, gridy, gridz, fields = ".dummy")
}
\arguments{
\item{gridx, gridy, gridz}{Grid coordinates in the three directions. The best
way to imput these is through the \code{seq()} function.}

\item{fields}{Column names to populate the grid. The columns are
automatically filled with \code{NA}s.}
}
\description{
A 3D grid with equally spaced points. Extends the \code{points3DDataFrame}
class.

A 3D grid with equally spaced points. Extends the \code{points3DDataFrame}
class.
}
\details{
Subsetting an object of this class will coerce the output to a
\code{points3DDataFrame} object.

The current implementation does not check if the points are
actually equally spaced in each direction. A future version will support
rotated grids.
}
\section{Slots}{

\describe{
\item{\code{dims}}{The number of grid points in each direction.}
}}

\seealso{
\code{\link{grid3DDataFrame-init}}
}
