% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines3DDataFrame.R
\docType{class}
\name{lines3DDataFrame-class}
\alias{lines3DDataFrame-class}
\alias{lines3DDataFrame}
\title{Drillhole database}
\description{
Drillhole database represented as line segments. Extends the
\code{spatial3DdataFrame} class.
}
\details{
The \code{"HOLEID"} column is used by this object's methods. Do not
remove or rename it.
}
\section{Slots}{

\describe{
\item{\code{coords}}{A \code{list} containing the coordinates for each data unit.}

\item{\code{data}}{A \code{data.frame} with each data unit's attributes.}

\item{\code{bbox}}{A \code{matrix} containing the coordinates of two opposing
edges of the object's bounding box.}
}}

\seealso{
\code{\link{lines3DDataFrame-init}},
\code{\link{spatial3DDataFrame-class}}
}
