% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines3DDataFrame.R
\name{lines3DDataFrame-init}
\alias{lines3DDataFrame-init}
\alias{lines3DDataFrame}
\title{Drillhole database}
\usage{
lines3DDataFrame(lines_list = NULL, df = NULL, collar = NULL,
  assay = NULL, survey = NULL, holeid = "HOLEID", from = "FROM",
  to = "TO", X = "X", Y = "Y", Z = "Z")
}
\arguments{
\item{lines_list}{A \code{list} containing the line segments' coordinates.}

\item{df}{A \code{data.frame} with the line segments' attributes.}

\item{collar, assay, survey}{Data frames with drillhole data.}

\item{holeid, from, to, X, Y, Z}{Column names from which to draw the parameters.}
}
\description{
Drillhole database represented as line segments.
}
\details{
There are two ways to build a \code{lines3DDataFrame} object:
either providing the data in the format that is seen in the final object
(less common) or by providing collar and assay data (more common).

In the first mode, both \code{lines_list} and \code{df} must be given.
\code{lines_list} is a \code{list} where each entry is a numeric vector
of length 6 containing the X, Y, and Z coordinates for both ends of a
line segment. \code{df} is a data frame containing the attributes for
each segment.

In the second mode, both \code{collar} and \code{assay} must be given.
\code{collar} contains the X, Y, and Z coordinates for each drillhole's
collar. \code{assay} contains the lithological description, chemical
analysis, etc. in intervals delimimted by the \code{from} and \code{to}
columns. \code{survey} contains the dipmeter measurements for each interval.
If it is missing, the holes are assumed vertical. The \code{holeid} column
must be present in all three data frames.
}
\seealso{
\code{\link{spatial3DDataFrame-class}},
\code{\link{lines3DDataFrame-class}}
}
