% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/grid3DDataFrame.R
\docType{methods}
\name{Make3DArray}
\alias{Make3DArray}
\alias{Make3DArray,grid3DDataFrame-method}
\title{Make3DArray}
\usage{
Make3DArray(object, ...)

\S4method{Make3DArray}{grid3DDataFrame}(object, value)
}
\arguments{
\item{value}{The column name to use. It can represent either a continuous or
a categorical variable.}
}
\value{
A \code{list} containing the following:
\describe{
  \item{value}{The 3D array itself.}
  \item{x, y, z}{The coordinates in each direction.}
}
}
\description{
Converts a regularly spaced 3D object to a 3D array.
}
