% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/points3DDataFrame.R,
%   R/lines3DDataFrame.R, R/grid3DDataFrame.R
\docType{methods}
\name{Pointify}
\alias{Pointify}
\alias{Pointify,points3DDataFrame-method}
\alias{Pointify,lines3DDataFrame-method}
\alias{Pointify,grid3DDataFrame-method}
\title{Conversion to point cloud}
\usage{
Pointify(object, ...)

\S4method{Pointify}{points3DDataFrame}(object)

\S4method{Pointify}{lines3DDataFrame}(object, locations = c(0.05, 0.5, 0.95),
  distance = F)

\S4method{Pointify}{grid3DDataFrame}(object)
}
\arguments{
\item{object}{The object being converted.}

\item{locations}{Relative locations along each line segment from which to
extract a point (between 0 and 1, inclusive).}

\item{distance}{Should the distance from the closest vertex be included?}
}
\value{
A \code{points3DDataFrame} object.
}
\description{
Converts any \code{spatial3DDataFrame} object to a \code{points3DDataFrame}.
Specific options depend of the object being converted.
}
\details{
If \code{distance = T}, the distance from the point to the closest
vertex in the original line segment is included in the returned object. The
values are stored in the \code{.dist} column.

In the current version, this method is only useful for
\code{lines3DDataFrame} objects. Future versions will add functionality for
block models and other kinds of objects.
}
