% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points3DDataFrame.R
\name{points3DDataFrame-class}
\alias{points3DDataFrame-class}
\alias{points3DDataFrame}
\title{points3DDataFrame-class}
\usage{
points3DDataFrame(coords, df)
}
\section{Slots}{

\describe{
\item{\code{coords}}{A \code{list} containing the coordinates for each data unit.}

\item{\code{data}}{A \code{data.frame} with each data unit's attributes.}

\item{\code{bbox}}{A \code{matrix} containing the coordinates of two opposing
edges of the object's bounding box.}
}}

\seealso{
\code{\link{spatial3DDataFrame-class}},
\code{\link{points3DDataFrame-init}}
}
