% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/GP.R, R/GP_geomod.R, R/SPGP.R
\docType{methods}
\name{Predict}
\alias{Predict}
\alias{Predict,GP-method}
\alias{Predict,GP_geomod-method}
\alias{Predict,SPGP-method}
\title{Predict}
\usage{
Predict(object, ...)

\S4method{Predict}{GP}(object, target, to = "value", output.var = T)

\S4method{Predict}{GP_geomod}(object, target, to = "value", output.ind = T,
  output.prob = T, use.unknown = T, Nsamp = 10000)

\S4method{Predict}{SPGP}(object, target, to = "value", output.var = T)
}
\arguments{
\item{target}{The \code{spatial3DDataFrame} object to receive the prediction.}

\item{to}{The name of the column in which to write the prediction. Will be
overwritten if it exists or created if not.}

\item{output.var}{Should the predictive variance be computed?}

\item{output.ind}{Return indicators for boundary drawing?}

\item{output.prob}{Return class probabilities?}

\item{use.unknown}{Include the unknown class in output?}

\item{Nsamp}{Number of samples used to estimate class probabilities.}
}
\value{
A 3D spatial object of the same class as \code{target}
containing the predictions.
}
\description{
Predictions at new spatial locations.
}
\details{
\code{GP} and \code{SPGP} objects return the predicted mean and
variance for each location in \code{target}. The sparse GP returns two
variances: \code{var_full} represents the total prediction uncertainty while
\code{var_cor} is the amplitude of variation of the underlying latent
function. The proportion between the two varies according to the distance
from the pseudo_inputs, and is given in the \code{quality} column.

The \code{GP_geomod} object will calculate an indicator and its variance
for each class at each location, which jointly form a multivariate normal
distribution of the true indicators (or log-transformed compositional
coordinates). The returned probabilities are actually the proportion of this
probability mass over the region in which each indicator is dominant. This
quantity is approximated by drawing \code{Nsamp} samples from the
distribution and computing the number of times each indicator is higher
than the others. The \code{use.unknown = F} option simply drops the
probability of the unknown class and re-normalizes the rest in order for
them to add to 1.
}
