% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/points3DDataFrame.R,
%   R/grid3DDataFrame.R
\docType{methods}
\name{SelectRegion}
\alias{SelectRegion}
\alias{SelectRegion,points3DDataFrame-method}
\alias{SelectRegion,grid3DDataFrame-method}
\title{SelectRegion}
\usage{
SelectRegion(object, ...)

\S4method{SelectRegion}{points3DDataFrame}(object, xmin = -Inf, xmax = Inf,
  ymin = -Inf, ymax = Inf, zmin = -Inf, zmax = Inf)

\S4method{SelectRegion}{grid3DDataFrame}(object, xmin = -Inf, xmax = Inf,
  ymin = -Inf, ymax = Inf, zmin = -Inf, zmax = Inf)
}
\arguments{
\item{xmin, xmax, ymin, ymax, zmin, zmax}{The limits of the box.}
}
\value{
A 3D object of the same class as the original.
}
\description{
Subsets a 3D object within a box specified by the user.
}
