% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/GP.R, R/SPGP.R
\docType{methods}
\name{Simulate}
\alias{Simulate}
\alias{Simulate,GP-method}
\alias{Simulate,SPGP-method}
\title{Geostatistical simulation}
\usage{
Simulate(object, ...)

\S4method{Simulate}{GP}(object, target, to = "value", Nsim,
  discount.noise = F)

\S4method{Simulate}{SPGP}(object, target, Nsim, to = "value",
  discount.noise = F, smooth = T, verbose = T, randnum = NULL)
}
\arguments{
\item{object}{A \code{GP} object.}

\item{target}{The \code{spatial3DDataFrame} object to receive the prediction.}

\item{to}{The name of the column in which to write the prediction. Will be
overwritten if it exists or created if not.}

\item{Nsim}{The desired number of simulations.}

\item{discount.noise}{Whether to sample from the GP latent function instead
of the noisy values.}

\item{smooth}{Whether to correct the output from the sparse GP.}

\item{verbose}{Display status while running?}
}
\value{
A 3D spatial object of the same class as \code{target}
containing the simulations.
}
\description{
Generates a number of realizations of the modeled random field.
}
\details{
Standard GP objects use the Cholesky method, which limits the number
of data to be used and the number of test points in which to simulate. SPGP
objects use a sequential method based on rank-one updates to the matrices
involved, which only limits the number of pseudo-inputs that can be used.

Due to the nature of the sparse approximation of the covariance matrices,
the output may appear noisy even if \code{discount.noise = T}. Set
\code{smooth = T} to apply a correction to the generated samples.
}
