% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial3DDataFrame.R
\docType{class}
\name{spatial3DDataFrame-class}
\alias{spatial3DDataFrame-class}
\alias{spatial3DDataFrame}
\title{Abstract class for 3D spatial objects}
\description{
Abstract class for 3D spatial objects
}
\details{
This class does not have a constructor method. It is used only to
define methods that are common to all of its subclasses.
}
\section{Slots}{

\describe{
\item{\code{coords}}{A \code{list} containing the coordinates for each data unit.
The contents of the \code{list} vary according to each object:
\itemize{
  \item For \code{points3DDataFrame} objects and its subclasses, a numeric
  vector of length 3;
  \item For \code{lines3DDataFrame} objects, a numeric vector of length 6,
  where the first 3 elements contain the coordinates of a line segment's
  start and the last 3 contain the coordinates of the segment's end.
}}

\item{\code{data}}{A \code{data.frame} with each data unit's attributes.}

\item{\code{bbox}}{A \code{matrix} containing the coordinates of two opposing
edges of the object's bounding box.}
}}

\seealso{
\code{\link{lines3DDataFrame-init}},
\code{\link{points3DDataFrame-init}}
}
