% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPGP.R
\name{SPGP-class}
\alias{SPGP-class}
\alias{SPGP}
\title{Sparse Gaussian Process}
\usage{
SPGP(data, model, value, mean = NULL, trend = NULL, pseudo_inputs,
  weights = NULL, force.interp = numeric(), reg.v = 1e-09,
  tangents = NULL, reg.t = 1e-12, pseudo_tangents = NULL,
  variational = T)
}
\description{
Implementation of the Sparse Gaussian Process model for 3D spatial
interpolation. Extends the \code{GP} class.
}
\section{Slots}{

\describe{
\item{\code{data}}{A \code{spatial3DDataFrame} object containing the necessary data.}

\item{\code{tangents}}{A \code{directions3DDataFrame} object containing structural
geology data. Most likely generated with the \code{GetLineDirections()}
or \code{GetPlaneDirections()} method.}

\item{\code{model}}{The covariance model. A \code{covarianceModel3D} object.}

\item{\code{mean}}{The global mean. Irrelevant if a trend is used.}

\item{\code{trend}}{The model's trend component. A formula in character format.}

\item{\code{beta}}{The trend coefficients.}

\item{\code{likelihood}}{The model's log-likelihood given the data.}

\item{\code{pre_comp}}{A \code{list} containing pre-computed values to speed up
predictions.}

\item{\code{pseudo_inputs}}{A \code{spatial3DDataFrame} object containing the coordinates
of the pseudo-inputs.}

\item{\code{pseudo_tangents}}{A \code{directions3DDataFrame} object containing the coordinates
of the pseudo-inputs for directional data.}

\item{\code{variational}}{A logical indicating if the model uses the variational
approach.}
}}

\references{
Snelson, E., Ghahramani, Z., 2006. Sparse Gaussian Processes using
Pseudo-inputs. Adv. Neural Inf. Process. Syst. 18 1257–1264.

Titsias, M., 2009. Variational Learning of Inducing Variables in Sparse
Gaussian Processes. Aistats 5, 567–574.

Bauer, M.S., van der Wilk, M., Rasmussen, C.E., 2016. Understanding
Probabilistic Sparse Gaussian Process Approximations. Adv. Neural Inf.
Process. Syst. 29.
}
\seealso{
\code{\link{SPGP-init}}, \code{\link{GP-class}}
}
