% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPGP.R
\name{SPGP-init}
\alias{SPGP-init}
\alias{SPGP}
\title{Sparse Gaussian Process}
\usage{
SPGP(data, model, value, mean = NULL, trend = NULL, pseudo_inputs,
  weights = NULL, force.interp = numeric(), reg.v = 1e-09,
  tangents = NULL, reg.t = 1e-12, pseudo_tangents = NULL,
  variational = T)
}
\arguments{
\item{data}{A \code{spatial3DDataFrame} object containing the data one
wishes to model.}

\item{model}{The covariance model. A \code{covarianceStructure3D} or a
\code{list} containing one or more such objects.}

\item{value}{The column name of the variable to be modeled. It is assumed
the column does not contain missing values.}

\item{mean}{The global mean. Irrelevant if a trend is provided.}

\item{trend}{The model's trend component. A formula in character format.}

\item{pseudo_inputs}{The desired number of pseudo-inputs (whose coordinates
will be sampled from the data), a matrix or data frame with their
coordinates, or a 3D object.}

\item{weights}{The importance of each data point in the model (a vector with
values between 0 and 1).}

\item{force.interp}{Indices of points that must be interpolated exactly.}

\item{reg.v}{Regularization to improve stability. A single value or a vector
with length matching the number of data points.}

\item{tangents}{A \code{directions3DDataFrame} object containing structural
geology data. Most likely generated with the \code{GetLineDirections()}
method.}

\item{reg.t}{Regularization for structural data. A single value or a vector
with length matching the number of structural data.}

\item{pseudo_tangents}{The desired number of pseudo-structural data (whose
coordinates will be sampled from the data) or a \code{directions3DDataFrame}.}

\item{variational}{Use the variational approach?}
}
\description{
Implementation of the Sparse Gaussian Process model for 3D spatial
interpolation. Extends the \code{GP} class.
}
\details{
This method builds a \code{SPGP} object with all the information
needed to make preditions at new data points.

\code{trend} must be a character string with a formula as a function of
uppercase X, Y, and Z. The most common is the linear trend,
\code{"~ X + Y + Z"}. For ordinary kriging, use \code{"~1"}. If neither
\code{trend} nor \code{mean} are given, it is assumed that the global mean
is the mean of the data values.

The SPGP works by compressing the information coming from all data into a
small number of pseudo-inputs. This way computational gains are obtained,
but the resulting model may pose difficulties for training.

Given the sparse nature of the model, the points specified in \code{force.interp}
may still not be interpolated exactly. The effects of tangent data may also be
diminished.

The variational model is cited in the literature as more stable and
less prone to overfitting. \code{variational = F} corresponds to the
Fully Independent Conditional (FIC) approach.
}
\references{
Snelson, E., Ghahramani, Z., 2006. Sparse Gaussian Processes using
Pseudo-inputs. Adv. Neural Inf. Process. Syst. 18 1257–1264.

Titsias, M., 2009. Variational Learning of Inducing Variables in Sparse
Gaussian Processes. Aistats 5, 567–574.

Bauer, M.S., van der Wilk, M., Rasmussen, C.E., 2016. Understanding
Probabilistic Sparse Gaussian Process Approximations. Adv. Neural Inf.
Process. Syst. 29.
}
\seealso{
\code{\link{SPGP-class}}, \code{\link{GP-class}}
}
