% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPGP_geomod.R
\docType{class}
\name{SPGP_geomod-class}
\alias{SPGP_geomod-class}
\alias{SPGP_geomod}
\title{Sparse Gaussian Process for Implicit Geological Modeling}
\description{
Implicit geological using Gaussian Processes and compositional relations
among the geological classes. Extends the \code{GP_geomod} class.
}
\section{Slots}{

\describe{
\item{\code{GPs}}{A \code{list} containing one \code{GP} object per class.}

\item{\code{params}}{A \code{list} containing modeling parameters such as the noise
variance and regularization factors.}
}}

\seealso{
\code{\link{SPGP_geomod-init}}
}
