% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPGP_geomod.R
\docType{methods}
\name{SPGP_geomod-init}
\alias{SPGP_geomod-init}
\title{Sparse Gaussian Process for Implicit Geological Modeling}
\usage{
\S4method{initialize}{SPGP_geomod}(.Object, data, value1, value2 = value1,
  model, nugget, tangents = NULL, pseudo_inputs, pseudo_tangents = NULL,
  variational = T, enforce.contacts = F, reg.v = 1e-09, reg.t = 1e-09)
}
\arguments{
\item{data}{A 3D spatial object.}

\item{value1}{A column name indicating the variable that contains the
geological classes.}

\item{value2}{Another column name, with possibly different class labels.}

\item{model}{The covariance model. A \code{covarianceStructure3D} or a
\code{list} containing one or more such objects.}

\item{nugget}{The model's nugget effect or noise variance.}

\item{tangents}{A \code{directions3DDataFrame} object containing structural
geology data. Most likely generated with the \code{GetLineDirections()}
method.}

\item{pseudo_inputs}{The desired number of pseudo-inputs (whose coordinates
will be sampled from the data) or a matrix or data frame with their
coordinates.}

\item{pseudo_tangents}{The desired number of pseudo-structural data (whose
coordinates will be sampled from the data) or a \code{directions3DDataFrame}.}

\item{variational}{Use the variational approach?}

\item{enforce.contacts}{Force the model to interpolate geological contacts?}

\item{reg.v}{Regularization to improve stability. A single value or a vector
with length matching the number of data points.}

\item{reg.t}{Regularization for structural data. A single value or a vector
with length matching the number of structural data.}
}
\description{
Implicit geological using Gaussian Processes and compositional relations
among the geological classes.
}
\details{
The role of this class is to manage multiple \code{SPGP} objects, one
per geological class in the data, that model an indicator variable or
"potential" for each class. A point in 3D space is assigned to the class
with the highest modeled potential. The class labels are defined as
\code{unique(c(data[[value1]], data[[value2]]))}.

The points at which \code{data[[value1]] != data[[value2]]} are considered
to lie exactly at a boundary between two geological classes.
If \code{enforce.contacts = T}
the \code{force.interp} option of the \code{GP} class is turned on in order
to make the geological boundaries pass through them.

The prediction provides an indicator value for each class at each location,
as well as the label of the most likely class. In locations away from the
data, the most likely class can be an extra, \code{"Unknown"} class. This
class is there to to ensure that the probabilities sum to 1 according to
compositional data principles, and also as a measure of model uncertainty.

The geological boundaries can be drawn by contouring each class's indicator
at the value 0.
}
\references{
[1] Gonçalves ÍG, Kumaira S, Guadagnin F.
A machine learning approach to the potential-field method for implicit
modeling of geological structures. Comput Geosci 2017;103:173–82.
doi:10.1016/j.cageo.2017.03.015.
}
\seealso{
\code{\link{SPGP}}, \code{\link{SPGP-init}}, \code{\link{Predict}},
\code{\link{Make3DArray}}
}
