% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/geostat3D.R
\docType{methods}
\name{TrendMatrix}
\alias{TrendMatrix}
\alias{TrendMatrix,points3DDataFrame,character-method}
\alias{TrendMatrix,directions3DDataFrame,character-method}
\title{Trend matrix}
\usage{
TrendMatrix(x, trend, ...)

\S4method{TrendMatrix}{points3DDataFrame,character}(x, trend)

\S4method{TrendMatrix}{directions3DDataFrame,character}(x, trend)
}
\arguments{
\item{x}{A 3D spatial object.}

\item{trend}{A character string convertable to a formula.}
}
\value{
A matrix with \code{nrow(x)} rows and number of columns according to
the specified trend formula.
}
\description{
Calculates a matrix containing the components of a trend function at the
specified data locations.
}
\details{
A call with a \code{directions3DDataFrame} as argument returns
the derivative of the trend function at the given locations. This is used in
situations involving derivative data.
}
\seealso{
\code{\link{CovarianceMatrix}}
}
