% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/SPGP.R
\docType{methods}
\name{Xval}
\alias{Xval}
\alias{Xval,SPGP-method}
\title{Cross-validation}
\usage{
Xval(object, ...)

\S4method{Xval}{SPGP}(object)
}
\arguments{
\item{object}{A \code{SPGP} object.}
}
\value{
A \code{list} with the following elements:
\describe{
  \item{mean}{The predictive mean for each data point.}
  \item{var}{The predictive variance.}
  \item{RMSE}{Root Mean Square Error of prediction, taking only the mean in
  consideration}
  \item{NRMSE}{Normalized RMSE: the squared difference is divided by the
  predictive variance.}
  \item{LPD}{Log Predictive Density.}
  \item{PLPD}{Penalized LPD: includes a penalty based on how well the full
  covariance is approximated at each data point.}
}
}
\description{
Performs cross-validation and calculates a number of performance metrics
}
\details{
The method performs leave-one-out cross-validation through
rank-one downdates to a covariance matrix.
}
