% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks3DDataFrame.R
\name{blocks3DDataFrame-class}
\alias{blocks3DDataFrame-class}
\alias{blocks3DDataFrame}
\title{3D block model}
\usage{
blocks3DDataFrame(corner, number, size, discretization = c(4, 4, 4),
  fields = ".dummy")
}
\description{
A 3D grid with adjacent blocks. Extends the \code{grid3DDataFrame}
class.
}
\details{
Subsetting an object of this class will coerce the output to a
\code{points3DDataFrame} object.
}
\section{Slots}{

\describe{
\item{\code{size}}{The block size in the three dimensions.}

\item{\code{discretization}}{The number of subdivisions for computation of
covariances.}
}}

\seealso{
\code{\link{blocks3DDataFrame-init}}
}
