% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/GP.R, R/GP_geomod.R
\docType{methods}
\name{Fit}
\alias{Fit}
\alias{Fit,GP-method}
\alias{Fit,GP_geomod-method}
\title{Fit}
\usage{
Fit(object, ...)

\S4method{Fit}{GP}(object, midrange = F, minrange = F, azimuth = F,
  dip = F, rake = F, nugget = F, power = F)

\S4method{Fit}{GP_geomod}(object, maxrange = T, midrange = F,
  minrange = F, azimuth = F, dip = F, rake = F, power = F)
}
\arguments{
\item{midrange, minrange}{Optimize on the covariance model's anisotropy?}

\item{azimuth, dip, rake}{Optimize on the covariance model's orientation?}

\item{nugget}{Optimize on nugget?}

\item{power}{Optimize on the power parameter (not relevant for all
covariances)?}
}
\value{
A \code{GP} object similar to \code{object}, with optimized
covariance parameters.
}
\description{
Training of a Gaussian Process model with a genetic algorithm.
}
\details{
This method uses a genetic algorithm to optimize the contribution
(or amplitude) and range of each covariance structure contained in the
\code{model} slot of \code{object}, as well as the parameters above, if
allowed. Optimization is done with respect to the object's log-likelihood.

By default the genetic algorithm will run for a maximum of 100 iterations or
20 iterations without improvement of the log-likelihood. If desired, this
method can be called multiple times to continue the optimization starting
from previous results.

In order to obtain reproducible results, set the seed before calling this
method.
}
