% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GP_geomod.R
\docType{class}
\name{GP_geomod-class}
\alias{GP_geomod-class}
\alias{GP_geomod}
\title{Gaussian Process for Implicit Geological Modeling}
\description{
Implicit geological using Gaussian Processes and compositional relations
among the geological classes.
}
\section{Slots}{

\describe{
\item{\code{GPs}}{A \code{list} containing one \code{GP} object per class.}

\item{\code{params}}{A \code{list} containing modeling parameters such as the noise
variance and regularization factors.}
}}

\seealso{
\code{\link{GP_geomod-init}}
}
