% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid3DDataFrame.R
\docType{methods}
\name{grid3DDataFrame-init}
\alias{grid3DDataFrame-init}
\title{3D regular point grid}
\usage{
\S4method{initialize}{grid3DDataFrame}(.Object, gridx, gridy, gridz,
  fields = ".dummy")
}
\arguments{
\item{gridx, gridy, gridz}{Grid coordinates in the three directions. The best
way to imput these is through the \code{seq()} function.}

\item{fields}{Column names to populate the grid. The columns are
automatically filled with \code{NA}s.}
}
\description{
A 3D grid with equally spaced points. Extends the \code{points3DDataFrame}
class.
}
\details{
The current implementation does not check if the points are
actually equally spaced in each direction. A future version will support
rotated grids.
}
