% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/GP.R, R/GP_geomod.R
\docType{methods}
\name{Predict}
\alias{Predict}
\alias{Predict,GP-method}
\alias{Predict,GP_geomod-method}
\title{Predict}
\usage{
Predict(object, ...)

\S4method{Predict}{GP}(object, target, to = "value", output.var = F)

\S4method{Predict}{GP_geomod}(object, target, to = "value")
}
\arguments{
\item{target}{The \code{spatial3DDataFrame} object to receive the prediction.}

\item{to}{The name of the column in which to write the prediction. Will be
overwritten if it exists or created if not.}

\item{output.var}{Should the predictive variance be computed?}
}
\value{
A 3D spatial object of the same class as \code{target}
containing the predictions.
}
\description{
Spatial interpolation at new locations.
}
