% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPGP_geomod.R
\name{SPGP_geomod-class}
\alias{SPGP_geomod-class}
\alias{SPGP_geomod}
\title{Sparse Gaussian Process for Implicit Geological Modeling}
\usage{
SPGP_geomod(data, value1, value2 = value1, model, nugget, tangents = NULL,
  pseudo_inputs, pseudo_tangents = NULL, variational = T,
  enforce.contacts = F, reg.v = 1e-09, reg.t = 1e-09)
}
\description{
Implicit geological using Gaussian Processes and compositional relations
among the geological classes. Extends the \code{GP_geomod} class.
}
\section{Slots}{

\describe{
\item{\code{GPs}}{A \code{list} containing one \code{GP} object per class.}

\item{\code{params}}{A \code{list} containing modeling parameters such as the noise
variance and regularization factors.}
}}

\seealso{
\code{\link{SPGP_geomod-init}}, \code{\link{GP_geomod-class}}
}
